drop 'path:' prefix from --override-input
`--override-input foo/bar` and `--override-input path:foo/bar` are
not identical. The latter copies foo/bar to the Nix store, which
can be painfully slow if it's hosted on a network filesystem and/or
if it contains a lot of small files (e.g. git blobs). The former
loads foo/bar from the filesystem directly.

Switching to the former greatly speeds up the Terraform plan phase.